requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
Django==4.2.2
djangorestframework==3.14.0
emoji==0.6.0
filelock==3.12.2
fsspec==2023.6.0
huggingface-hub==0.15.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.25.0
packaging==23.1
pyre-extensions==0.0.29
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
sqlparse==0.4.4
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.30.2
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
